+++ core-java-modules/core-java-strings/src/test/java/com/baeldung/stringbuffer/ConcatPerformanceTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.concatperformance;
+package com.baeldung.stringbuffer;
 import com.baeldung.stringbuffer.ConcatPerformance;
 
 import org.junit.Test;
