+++ core-java-modules/core-java-strings/src/test/java/com/baeldung/stringbuffer/HashCodeTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.hashcodecompare;
+package com.baeldung.stringbuffer;
 import com.baeldung.stringbuffer.HashCode;
 
 import org.junit.Test;
